# Precalculate Non-zero divisors for lak