Simplify division step in Lak calculations so that we don't need to check for divide by zero.