@Melirius 2024-06-11
It is OK, do you want to change division later to division-by-multiplication here?