@mcroomp 2024-06-13
Not sure the division-by-multiply will fly because of the i16 cast to determine the sign. It messes up a bunch of potential optimizations.